Fix Geckodriver tests
