+++ .gitignore
@@ -6,3 +6,4 @@ composer.lock
 composer.phar
 .idea/
 .phpunit.result.cache
+phpunit-results.xml
